Rearrange + Dev Decoding Cont
- functions/decoding_funcs.py: removed plotting functions to their own file.
- functions/dependent_bayes_analysis.py: update function call to correct new file for plotting.
- functions/dependent_bayes_deviation_analysis.py: added parameters and updated function calls
- functions/dependent_decoding_funcs.py: plot renaming, unnecessary variable cleanup, added z-score handling, added decoding and storage of pre- and post- intervals surrounding deviation events.
- functions/plot_decoding_funcs.py: new file for plotting functions associated with sliding decoder.
- functions/plot_dev_decoding_funcs.py: new file for plotting functions associated with deviation decoder.
- utils/test_support.py: testing changes.
