+++ functions/dependent_decoding_funcs.py
@@ -402,7 +402,7 @@ def decode_epochs(tastant_fr_dist, segment_spike_times, post_taste_dt,
                 
                 # Pass inputs to parallel computation on probabilities
                 inputs = zip(list_tb_fr, itertools.repeat(num_tastes),
-                              itertools.repeat(all_taste_gmm), itertools.repeat(p_taste))
+                              itertools.repeat(all_taste_gmm), itertools.repeat(p_taste_train_counts))
                 tic = time.time()
                 pool = Pool(4)
                 tb_decode_prob = pool.map(
@@ -481,9 +481,9 @@ def decode_epochs(tastant_fr_dist, segment_spike_times, post_taste_dt,
             plt.legend(loc='right')
             plt.ylabel('Decoding Fraction')
             plt.xlabel('Time (min)')
-            plt.title('Segment ' + str(s_i))
-            f1.savefig(os.path.join(seg_decode_save_dir,'segment_' + str(s_i) + '.png'))
-            f1.savefig(os.path.join(seg_decode_save_dir, 'segment_' + str(s_i) + '.svg'))
+            plt.title((' ').join(segment_names[s_i].split('-')))
+            f1.savefig(os.path.join(seg_decode_save_dir,segment_names[s_i] + '.png'))
+            f1.savefig(os.path.join(seg_decode_save_dir,segment_names[s_i] + '.svg'))
             plt.close(f1)
             # Imshow
             f2 = plt.figure()
@@ -497,26 +497,24 @@ def decode_epochs(tastant_fr_dist, segment_spike_times, post_taste_dt,
             plt.yticks(y_ticks, dig_in_names)
             plt.ylabel('Decoding Fraction')
             plt.xlabel('Time (min)')
-            plt.title('Segment ' + str(s_i))
-            f2.savefig(os.path.join(seg_decode_save_dir,'segment_' + str(s_i) + '_im.png'))
-            f2.savefig(os.path.join(seg_decode_save_dir,'segment_' + str(s_i) + '_im.svg'))
+            plt.title((' ').join(segment_names[s_i].split('-')))
+            f2.savefig(os.path.join(seg_decode_save_dir,segment_names[s_i] + '_im.png'))
+            f2.savefig(os.path.join(seg_decode_save_dir,segment_names[s_i] + '_im.svg'))
             plt.close(f2)
             # Fraction of occurrences
             f3 = plt.figure()
             plt.pie(np.sum(seg_decode_epoch_taste_bin, 1)/np.sum(seg_decode_epoch_taste_bin),
-                    labels=['water', 'saccharin', 'none'], autopct='%1.1f%%', pctdistance=1.5,
+                    labels=dig_in_names, autopct='%1.1f%%', pctdistance=1.5,
                     colors=taste_colors)
-            plt.title('Segment ' + str(s_i))
-            f3.savefig(os.path.join(seg_decode_save_dir,'segment_' + str(s_i) + '_pie.png'))
-            f3.savefig(os.path.join(seg_decode_save_dir,'segment_' + str(s_i) + '_pie.svg'))
+            plt.title((' ').join(segment_names[s_i].split('-')))
+            f3.savefig(os.path.join(seg_decode_save_dir,segment_names[s_i] + '_pie.png'))
+            f3.savefig(os.path.join(seg_decode_save_dir,segment_names[s_i] + '_pie.svg'))
             plt.close(f3)
             
-def decode_deviations_epochs(tastant_fr_dist, segment_spike_times, post_taste_dt,
-                  e_skip_dt, e_len_dt, dig_in_names, segment_times,
-                  segment_names, start_dig_in_times, taste_num_deliv,
-                  segment_dev_rasters, segment_dev_times, taste_select_epoch,
-                  max_hz, save_dir, neuron_count_thresh, decode_prob_cutoff, 
-                  z_score = False, epochs_to_analyze=[], segments_to_analyze=[]):
+def decode_deviations_epochs(tastant_fr_dist, segment_spike_times, dig_in_names, 
+                  segment_times, segment_names, start_dig_in_times, taste_num_deliv,
+                  segment_dev_times, segment_dev_fr_vecs, taste_select_epoch, bin_dt, 
+                  save_dir, z_score = True, epochs_to_analyze=[], segments_to_analyze=[]):
     """Decode taste from epoch-specific firing rates"""
     print('\t\tRunning GMM Decoder')
     # Variables
@@ -524,13 +522,11 @@ def decode_deviations_epochs(tastant_fr_dist, segment_spike_times, post_taste_dt
     num_neur = len(segment_spike_times[0])
     num_cp = len(tastant_fr_dist[0][0])
     num_segments = len(segment_spike_times)
-    p_taste = taste_num_deliv/np.sum(taste_num_deliv)  # P(taste)
+    #p_taste = taste_num_deliv/np.sum(taste_num_deliv)  # P(taste)
     cmap = colormaps['jet']
     taste_colors = cmap(np.linspace(0, 1, num_tastes))
-    half_decode_bin_dt = np.ceil(e_len_dt/2).astype('int')
-    cmap = colormaps['cividis']
-    epoch_colors = cmap(np.linspace(0, 1, len(epochs_to_analyze)))
-
+    dev_buffer = 50
+    
     if len(epochs_to_analyze) == 0:
         epochs_to_analyze = np.arange(num_cp)
     if len(segments_to_analyze) == 0:
@@ -602,20 +598,15 @@ def decode_deviations_epochs(tastant_fr_dist, segment_spike_times, post_taste_dt
             # Insert here a line of fitting the Gamma-MM
             all_taste_gmm[t_i] = gm
 
-        # Segment-by-segment use full taste decoding times to zoom in and test
-        #	epoch-specific and smaller interval
-        for s_i in segments_to_analyze:
+        # Segment-by-segment use deviation rasters and times to zoom in and test
+        #	epoch-specific decoding of tastes. Add decoding of 50 ms on either
+        #	side of the deviation event as well for context decoding.
+        for seg_ind, s_i in enumerate(segments_to_analyze):
             # Get segment variables
             seg_start = segment_times[s_i]
             seg_end = segment_times[s_i+1]
             seg_len = segment_times[s_i+1] - segment_times[s_i]  # in dt = ms
-            new_time_bins = np.arange(
-                seg_start+half_decode_bin_dt, seg_end-half_decode_bin_dt, e_skip_dt)
-            # Now pull epoch-specific probabilities
-            seg_decode_epoch_prob = np.zeros((num_tastes, seg_len))
-            # Start with assumption of "none" taste at all times
-            seg_decode_epoch_prob[-1, :] = 1
-            # Binerize Spike Times
+            # Binerize Segment Spike Times
             segment_spike_times_s_i = segment_spike_times[s_i]
             segment_spike_times_s_i_bin = np.zeros((num_neur, seg_len+1))
             for n_i in taste_select_neur:
@@ -623,154 +614,150 @@ def decode_deviations_epochs(tastant_fr_dist, segment_spike_times, post_taste_dt
                     segment_spike_times_s_i[n_i] - seg_start).astype('int')
                 segment_spike_times_s_i_bin[n_i, n_i_spike_times] = 1
             
+            if z_score == True:
+                # Calculate mean and std of binned segment spikes for z-scoring
+                z_time_bins = np.arange(0,seg_len-bin_dt,bin_dt)
+                seg_fr = np.zeros((num_neur,len(z_time_bins))) #Hz
+                for bdt_i, bdt in enumerate(z_time_bins):
+                    seg_fr[:,bdt_i] = np.sum(segment_spike_times_s_i_bin[:,bdt:bdt+bin_dt],1)/(bin_dt/1000)
+                mean_fr = np.nanmean(seg_fr,1)
+                std_fr = np.nanstd(seg_fr,1)
+            
             # Grab neuron firing rates in sliding bins
             try:
-                seg_decode_epoch_prob = np.load(
-                    os.path.join(epoch_decode_save_dir,'segment_' + str(s_i) + '.npy'))
-                tb_fr = np.load(os.path.join(epoch_decode_save_dir,
-                                'segment_' + str(s_i) + '_tb_fr.npy'))
+                dev_decode_array = np.load(
+                    os.path.join(epoch_decode_save_dir,'segment_' + str(s_i) + \
+                                 '_deviations.npy'))
+                pre_dev_decode_array = np.load(
+                    os.path.join(epoch_decode_save_dir,'segment_' + str(s_i) + \
+                                 '_pre_deviations.npy'))
+                post_dev_decode_array = np.load(
+                    os.path.join(epoch_decode_save_dir,'segment_' + str(s_i) + \
+                                 '_post_deviations.npy'))
                 print('\t\t\t\tSegment ' + str(s_i) + ' Previously Decoded')
             except:
-                print('\t\t\t\tDecoding Segment ' + str(s_i))
-                # Perform parallel computation for each time bin
-                print('\t\t\t\t\tCalculate firing rates for time bins')
-                try:
-                    tb_fr = np.load(os.path.join(epoch_decode_save_dir,
-                                    'segment_' + str(s_i) + '_tb_fr.npy'))
-                except:
-                    if len(new_time_bins) > 1:
-                        tb_fr = np.zeros((num_neur, len(new_time_bins)))
-                        for tb_i, tb in enumerate(new_time_bins):
-                            tb_fr[:, tb_i] = np.sum(
-                                segment_spike_times_s_i_bin[:, tb-seg_start-half_decode_bin_dt:tb+half_decode_bin_dt-seg_start], 1)/(int(2*half_decode_bin_dt)/1000)
-                        np.save(os.path.join(epoch_decode_save_dir,'segment_' +
-                                str(s_i) + '_tb_fr.npy'), tb_fr)
-                    else:
-                        tb_fr = np.expand_dims(np.sum(segment_spike_times_s_i_bin,1)/((seg_len+1)/1000),1)
-                    del tb_i, tb
-                #Z-scoring
-                if z_score == True:
-                    #Calculate mean and std for binned spikes
-                    tb_fr_mean = np.mean(tb_fr,1)
-                    tb_fr_std = np.std(tb_fr,1)
-                    #Convert tb_fr to z-score
-                    tb_fr = np.divide(tb_fr - np.expand_dims(tb_fr_mean,1),np.expand_dims(tb_fr_std,1))
+                print('\t\t\t\tDecoding Segment ' + str(s_i) + ' Deviations')
+                
+                seg_dev_fr_mat = np.array(segment_dev_fr_vecs[seg_ind]).T
+                seg_dev_times = segment_dev_times[seg_ind]
+                _, num_dev = np.shape(seg_dev_times)
+                
+                #Pull pre-dev bin frs
+                pre_dev_fr_mat = []
+                for dev_i in range(num_dev):
+                    pre_dev_end = seg_dev_times[0,dev_i]
+                    pre_dev_start = np.max(pre_dev_end-dev_buffer,0)
+                    pre_dev_len = pre_dev_end - pre_dev_start
+                    pre_dev_bin = segment_spike_times_s_i_bin[:,pre_dev_start:pre_dev_end]
+                    pre_dev_fr = np.sum(pre_dev_bin,1)/(pre_dev_len/1000)
+                    if z_score == True:
+                        pre_dev_fr = (pre_dev_fr - mean_fr)/std_fr
+                    pre_dev_fr_mat.append(list(pre_dev_fr))
+                
+                #Pull post-dev bin frs
+                post_dev_fr_mat = []
+                for dev_i in range(num_dev):
+                    post_dev_start = seg_dev_times[1,dev_i]
+                    post_dev_end = np.min(pre_dev_end+dev_buffer,seg_len)
+                    post_dev_len = post_dev_end - post_dev_start
+                    post_dev_bin = segment_spike_times_s_i_bin[:,post_dev_start:post_dev_end]
+                    post_dev_fr = np.sum(post_dev_bin,1)/(post_dev_len/1000)
+                    if z_score == True:
+                        post_dev_fr = (post_dev_fr - mean_fr)/std_fr
+                    post_dev_fr_mat.append(list(post_dev_fr))
+                
                 #Converting to list for parallel processing
                 if np.min(true_taste_train_data) >= 0:    
-                    tb_fr_pca = pca_reduce.transform(tb_fr.T)
-                    list_tb_fr = list(tb_fr_pca)
+                    dev_fr_pca = pca_reduce.transform(seg_dev_fr_mat.T)
+                    list_dev_fr = list(dev_fr_pca)
                 else:
-                    list_tb_fr = list(tb_fr)
-                del tb_fr
+                    list_dev_fr = list(seg_dev_fr_mat.T)
+                del seg_dev_fr_mat
                 
                 # Pass inputs to parallel computation on probabilities
-                inputs = zip(list_tb_fr, itertools.repeat(num_tastes),
-                              itertools.repeat(all_taste_gmm), itertools.repeat(p_taste))
                 tic = time.time()
+                #Deviation Bins
+                inputs = zip(list_dev_fr, itertools.repeat(num_tastes),
+                              itertools.repeat(all_taste_gmm), itertools.repeat(p_taste_train_counts))
                 pool = Pool(4)
-                tb_decode_prob = pool.map(
+                dev_decode_prob = pool.map(
+                    dp.segment_taste_decode_dependent_parallelized, inputs)
+                pool.close()
+                #Pre-Deviation Bins
+                inputs = zip(pre_dev_fr_mat, itertools.repeat(num_tastes),
+                              itertools.repeat(all_taste_gmm), itertools.repeat(p_taste_train_counts))
+                pool = Pool(4)
+                pre_dev_decode_prob = pool.map(
+                    dp.segment_taste_decode_dependent_parallelized, inputs)
+                pool.close()
+                #Post-Deviation Bins
+                inputs = zip(post_dev_fr_mat, itertools.repeat(num_tastes),
+                              itertools.repeat(all_taste_gmm), itertools.repeat(p_taste_train_counts))
+                pool = Pool(4)
+                post_dev_decode_prob = pool.map(
                     dp.segment_taste_decode_dependent_parallelized, inputs)
                 pool.close()
                 toc = time.time()
                 print('\t\t\t\t\tTime to decode = ' +
                       str(np.round((toc-tic)/60, 2)) + ' (min)')
-                tb_decode_array = np.squeeze(np.array(tb_decode_prob)).T
-                # The whole skip interval should have the same decode probability
-                for s_dt in range(e_skip_dt):
-                    seg_decode_epoch_prob[:, new_time_bins -
-                                          seg_start + s_dt] = tb_decode_array
+                dev_decode_array = np.squeeze(np.array(dev_decode_prob)).T
+                pre_dev_decode_array = np.squeeze(np.array(pre_dev_decode_prob)).T
+                post_dev_decode_array = np.squeeze(np.array(post_dev_decode_prob)).T
                 # Save decoding probabilities
                 np.save(os.path.join(epoch_decode_save_dir,'segment_' +
-                        str(s_i) + '.npy'), seg_decode_epoch_prob)
+                        str(s_i) + '_deviations.npy'), dev_decode_array)
+                np.save(os.path.join(epoch_decode_save_dir,'segment_' +
+                        str(s_i) + '_pre_deviations.npy'), pre_dev_decode_array)
+                np.save(os.path.join(epoch_decode_save_dir,'segment_' +
+                        str(s_i) + '_post_deviations.npy'), post_dev_decode_array)
             # Create plots
             seg_decode_save_dir = os.path.join(epoch_decode_save_dir,
                 'segment_' + str(s_i) + '/')
             if not os.path.isdir(seg_decode_save_dir):
                 os.mkdir(seg_decode_save_dir)
 
-            seg_decode_epoch_prob_nonan = np.zeros(
-                np.shape(seg_decode_epoch_prob))
-            seg_decode_epoch_prob_nonan[:] = seg_decode_epoch_prob[:]
-            seg_decode_epoch_prob_nonan[np.isnan(
-                seg_decode_epoch_prob_nonan)] = 0
-            seg_decode_missing = np.where(np.sum(seg_decode_epoch_prob_nonan,0) == 0)[0]
-            seg_decode_epoch_prob_nonan[-1,seg_decode_missing] = 1 #Make it a no-taste timepoint
-            seg_decode_epoch_taste_ind = np.argmax(seg_decode_epoch_prob, 0)
-            
-            # Updated decoding based on threshold
-            seg_decode_epoch_taste_bin = np.zeros(
-                np.shape(seg_decode_epoch_prob))
+            _, num_dev = np.shape(dev_decode_array)
+            dev_decode_epoch_taste_ind = np.argmax(dev_decode_array, 0)
+            dev_decode_bin = np.zeros((num_tastes,num_dev))
             for t_i in range(num_tastes):
-                if t_i < num_tastes-1:
-                    taste_max_bin = (seg_decode_epoch_taste_ind == t_i).astype('int')
-                    taste_bin_cutoff = (seg_decode_epoch_prob[t_i,:] >= decode_prob_cutoff).astype('int')
-                    taste_bin = taste_max_bin*taste_bin_cutoff
-                    # To ensure starts and ends of bins align
-                    taste_bin[0] = 0
-                    taste_bin[-1] = 0
-                    # Calculate decoding periods
-                    diff_decoded_taste = np.diff(taste_bin)
-                    start_decoded = np.where(diff_decoded_taste == 1)[0] + 1
-                    end_decoded = np.where(diff_decoded_taste == -1)[0] + 1
-                    num_decoded = len(start_decoded)
-                    # Calculate number of neurons in each period
-                    num_neur_decoded = np.zeros(num_decoded)
-                    for nd_i in range(num_decoded):
-                        d_start = start_decoded[nd_i]
-                        d_end = end_decoded[nd_i]
-                        for n_i in range(num_neur):
-                            if len(np.where(segment_spike_times_s_i_bin[n_i, d_start:d_end])[0]) > 0:
-                                num_neur_decoded[nd_i] += 1
-                    # Now cut at threshold and only keep matching decoded intervals
-                    decode_ind = np.where(
-                        num_neur_decoded > neuron_count_thresh)[0]
-                    for db in decode_ind:
-                        s_db = start_decoded[db]
-                        e_db = end_decoded[db]
-                        seg_decode_epoch_taste_bin[t_i, s_db:e_db] = 1
-                else: #No taste everywhere the other taste is not marked
-                    where_decoded_taste = (np.sum(seg_decode_epoch_taste_bin,0) > 0).astype('int')
-                    where_not_decoded_taste = np.ones(len(where_decoded_taste)) - where_decoded_taste
-                    seg_decode_epoch_taste_bin[t_i,:] = where_not_decoded_taste
+                dev_decode_bin[t_i,np.where(dev_decode_epoch_taste_ind == t_i)[0]] = 1
+            
             # Line plot
             f1 = plt.figure()
             for t_i in range(num_tastes):
-                plt.plot(np.arange(seg_start, seg_end)/1000/60, \
-                         seg_decode_epoch_prob_nonan[t_i,:], \
+                plt.plot(np.arange(num_dev), dev_decode_array[t_i,:], \
                              color=taste_colors[t_i,:], label=dig_in_names[t_i])
-                plt.fill_between(np.arange(seg_start, seg_end)/1000/60, \
-                                 seg_decode_epoch_taste_bin[t_i, :], alpha=0.2 \
-                                 ,color=taste_colors[t_i,:], label='_')
+                plt.fill_between(np.arange(num_dev), dev_decode_array[t_i, :], \
+                                 alpha=0.2, color=taste_colors[t_i,:], label='_')
             plt.legend(loc='right')
             plt.ylabel('Decoding Fraction')
-            plt.xlabel('Time (min)')
-            plt.title('Segment ' + str(s_i))
-            f1.savefig(os.path.join(seg_decode_save_dir,'segment_' + str(s_i) + '.png'))
-            f1.savefig(os.path.join(seg_decode_save_dir, 'segment_' + str(s_i) + '.svg'))
+            plt.xlabel('Deviation Index')
+            plt.title((' ').join(segment_names[s_i].split('-')))
+            f1.savefig(os.path.join(seg_decode_save_dir, segment_names[s_i] + '.png'))
+            f1.savefig(os.path.join(seg_decode_save_dir, segment_names[s_i] + '.svg'))
             plt.close(f1)
             # Imshow
             f2 = plt.figure()
-            plt.imshow(seg_decode_epoch_prob_nonan,
+            plt.imshow(dev_decode_array,
                        aspect='auto', interpolation='none')
-            x_ticks = np.ceil(np.linspace(
-                0, len(new_time_bins)-1, 10)).astype('int')
-            x_tick_labels = np.round(new_time_bins[x_ticks]/1000/60, 2)
+            x_ticks = np.ceil(np.linspace(0, num_dev-1, 10)).astype('int')
+            x_tick_labels = np.round(np.arange(num_dev)[x_ticks], 2)
             plt.xticks(x_ticks, x_tick_labels)
             y_ticks = np.arange(len(dig_in_names))
             plt.yticks(y_ticks, dig_in_names)
             plt.ylabel('Decoding Fraction')
-            plt.xlabel('Time (min)')
-            plt.title('Segment ' + str(s_i))
-            f2.savefig(os.path.join(seg_decode_save_dir,'segment_' + str(s_i) + '_im.png'))
-            f2.savefig(os.path.join(seg_decode_save_dir,'segment_' + str(s_i) + '_im.svg'))
+            plt.xlabel('Deviation Index')
+            plt.title((' ').join(segment_names[s_i].split('-')))
+            f2.savefig(os.path.join(seg_decode_save_dir,segment_names[s_i] + '_im.png'))
+            f2.savefig(os.path.join(seg_decode_save_dir,segment_names[s_i] + '_im.svg'))
             plt.close(f2)
             # Fraction of occurrences
             f3 = plt.figure()
-            plt.pie(np.sum(seg_decode_epoch_taste_bin, 1)/np.sum(seg_decode_epoch_taste_bin),
-                    labels=['water', 'saccharin', 'none'], autopct='%1.1f%%', pctdistance=1.5,
+            plt.pie(np.sum(dev_decode_bin, 1)/num_dev,
+                    labels=dig_in_names, autopct='%1.1f%%', pctdistance=1.5,
                     colors=taste_colors)
-            plt.title('Segment ' + str(s_i))
-            f3.savefig(os.path.join(seg_decode_save_dir,'segment_' + str(s_i) + '_pie.png'))
-            f3.savefig(os.path.join(seg_decode_save_dir,'segment_' + str(s_i) + '_pie.svg'))
+            plt.title((' ').join(segment_names[s_i].split('-')))
+            f3.savefig(os.path.join(seg_decode_save_dir,segment_names[s_i] + '_pie.png'))
+            f3.savefig(os.path.join(seg_decode_save_dir,segment_names[s_i] + '_pie.svg'))
             plt.close(f3)
